updated data model to allow helices to be nonconsecutive
This is a breaking change because it changes the type of DNADesign.helices from List[Helix] to Map[int, Helix]. The JSON representation hasn't changed; now it simply allows an extra field called "idx" in each Helix. Most code should still work, except for looping over helices. Before that could be done with `for h in design.helices`. Now it must be done with `for h in design.helices.values()`. Similarly, `for i,h in enumerate(design.helices)` should now be `for i,h in design.helices.items()`.