Encoder indent suppression: str.replace → re.sub (fixes #170)
Using re.sub (re is already imported), along with a replacement
function that pulls from the replacement map, avoids repeated string
replacements within Python and significantly speeds up encoding for
large files.  Note that I think this will raise a KeyError if the user has
something with a name fitting "@@(\d+)@@" where the number is larger
than the largest unique_id, and will corrupt the file if it is not.
However, this is similar to the current situation.